requirements.txt
asgiref==3.7.2
Django==5.0.1
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2023.4
pilkit
django-imagekit
ultralytics
pandas
pymysql
sqlalchemy
mysqlclient
django-extensions
